Move iMapGenerator to engine
